# server/docs/whoop-openapi.json
{
  "openapi": "3.0.1",
  "info": { "title": "WHOOP API" },
  "servers": [{ "url": "https://api.prod.whoop.com/developer" }],
...
  "paths": {
...
    "/v2/cycle": {
      "get": {
        "tags": ["Cycle"], "operationId": "getCycleCollection",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "maximum": 25, "type": "integer", "format": "int32", "default": 10 } },
          { "name": "start", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "end", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "nextToken", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Successful request" }, "400": { "description": "Client error" }, "401": { "description": "Invalid authorization" }, "429": { "description": "Rate limiting" }, "500": { "description": "Server error" } },
        "security": [{ "OAuth": ["read:cycles"] }]
      }
    },
...
    "/v2/recovery": {
      "get": {
        "tags": ["Recovery"], "operationId": "getRecoveryCollection",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "maximum": 25, "type": "integer", "format": "int32", "default": 10 } },
          { "name": "start", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "end", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "nextToken", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Successful request" }, "400": { "description": "Client error" }, "401": { "description": "Invalid authorization" }, "429": { "description": "Rate limiting" }, "500": { "description": "Server error" } },
        "security": [{ "OAuth": ["read:recovery"] }]
      }
    },
...
    "/v2/activity/sleep": {
      "get": {
        "tags": ["Sleep"], "operationId": "getSleepCollection",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "maximum": 25, "type": "integer", "format": "int32", "default": 10 } },
          { "name": "start", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "end", "in": "query", "schema": { "type": "string", "format": "date-time" } },
          { "name": "nextToken", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Successful request" }, "400": { "description": "Client error" }, "401": { "description": "Invalid authorization" }, "429": { "description": "Rate limiting" }, "500": { "description": "Server error" } },
        "security": [{ "OAuth": ["read:sleep"] }]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "OAuth": { "type": "oauth2", "flows": { "authorizationCode": { "authorizationUrl": "https://api.prod.whoop.com/oauth/oauth2/auth", "tokenUrl": "https://api.prod.whoop.com/oauth/oauth2/token", "scopes": {
        "read:recovery": "Read Recovery data", "read:cycles": "Read cycles data", "read:workout": "Read workout data", "read:sleep": "Read sleep data", "read:profile": "Read profile data", "read:body_measurement": "Read body measurements data"
      } } } },
      "Trusted Partner": { "type": "oauth2", "flows": { "clientCredentials": { "tokenUrl": "https://api.prod.whoop.com/developer/v2/partner/token", "scopes": { "whoop-partner/token": "Read service requests and upload results." } } } }
    }
  }
}